+++ docs/CHANGELOG.rst
@@ -5,6 +5,11 @@ Change Log
 All notable changes to this project are documented in this file.
 
 
+==========
+Unreleased
+==========
+
+
 ===================
 10.0.0 - 2019-05-08
 ===================